+++ src/earthkit/regrid/data/numpy.py
@@ -1,0 +1,31 @@
+# (C) Copyright 2023 ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+#
+
+import logging
+
+from .handler import DataHandler
+
+LOG = logging.getLogger(__name__)
+
+
+class NumpyDataHandler(DataHandler):
+    @staticmethod
+    def match(values):
+        import numpy as np
+
+        return isinstance(values, np.ndarray)
+
+    def regrid(self, values, **kwargs):
+        in_grid = kwargs.pop("in_grid")
+        out_grid = kwargs.pop("out_grid")
+        backend = self.backend_from_kwargs(kwargs)
+        return backend.regrid(values, in_grid, out_grid, **kwargs)
+
+
+handler = NumpyDataHandler()
