+++ tests/regrid_backends/test_matrix_xarray.py
@@ -1,0 +1,13 @@
+# (C) Copyright 2023 ECMWF.
+#
+# This software is licensed under the terms of the Apache Licence Version 2.0
+# which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+# In applying this licence, ECMWF does not waive the privileges and immunities
+# granted to it by virtue of its status as an intergovernmental organisation
+# nor does it submit to any jurisdiction.
+
+
+from earthkit.regrid.utils.testing import NO_EKD  # noqa: E402
+
+if not NO_EKD:
+    from earthkit.data import from_source  # noqa
